+++ packages/api-client/src/main/team/legalhold/LegalHoldMemberStatus.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2019 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,7 +17,8 @@
  *
  */
 
-export enum ReactionType {
-  LIKE = '❤️',
-  NONE = '',
+export enum LegalHoldMemberStatus {
+  DISABLED = 'disabled',
+  ENABLED = 'enabled',
+  PENDING = 'pending',
 }