+++ packages/core/src/main/conversation/message/MessageBuilder.ts
@@ -19,7 +19,7 @@
 
 import {APIClient} from '@wireapp/api-client';
 import {ClientAction, Confirmation} from '@wireapp/protocol-messaging';
-import {AbortReason, PayloadBundleState, PayloadBundleType, ReactionType} from '..';
+import {AbortReason, PayloadBundleState, PayloadBundleType} from '..';
 import {AssetService} from '../AssetService';
 import {
   ClientActionContent,
@@ -33,6 +33,7 @@ import {
   ImageAssetContent,
   ImageContent,
   KnockContent,
+  LegalHoldStatus,
   LinkPreviewContent,
   LinkPreviewUploadedContent,
   LocationContent,
@@ -63,7 +64,7 @@ class MessageBuilder {
     conversationId: string,
     newMessageText: string,
     originalMessageId: string,
-    messageId: string = MessageBuilder.createId()
+    messageId = MessageBuilder.createId()
   ): TextContentBuilder {
     const content: EditedTextContent = {
       originalMessageId,
@@ -87,7 +88,8 @@ class MessageBuilder {
     conversationId: string,
     file: FileContent,
     originalMessageId: string,
-    expectsReadConfirmation?: boolean
+    expectsReadConfirmation?: boolean,
+    legalHoldStatus?: LegalHoldStatus
   ): Promise<FileAssetMessage> {
     const imageAsset = await this.assetService.uploadFileAsset(file);
 
@@ -95,6 +97,7 @@ class MessageBuilder {
       asset: imageAsset,
       expectsReadConfirmation,
       file,
+      legalHoldStatus,
     };
 
     return {
@@ -111,11 +114,13 @@ class MessageBuilder {
   public createFileMetadata(
     conversationId: string,
     metaData: FileMetaDataContent,
+    messageId = MessageBuilder.createId(),
     expectsReadConfirmation?: boolean,
-    messageId: string = MessageBuilder.createId()
+    legalHoldStatus?: LegalHoldStatus
   ): FileAssetMetaDataMessage {
     const content: FileAssetMetaDataContent = {
       expectsReadConfirmation,
+      legalHoldStatus,
       metaData,
     };
 
@@ -134,10 +139,12 @@ class MessageBuilder {
     conversationId: string,
     reason: AbortReason,
     originalMessageId: string,
-    expectsReadConfirmation?: boolean
+    expectsReadConfirmation?: boolean,
+    legalHoldStatus?: LegalHoldStatus
   ): Promise<FileAssetAbortMessage> {
     const content: FileAssetAbortContent = {
       expectsReadConfirmation,
+      legalHoldStatus,
       reason,
     };
 
@@ -155,8 +162,9 @@ class MessageBuilder {
   public async createImage(
     conversationId: string,
     image: ImageContent,
+    messageId = MessageBuilder.createId(),
     expectsReadConfirmation?: boolean,
-    messageId: string = MessageBuilder.createId()
+    legalHoldStatus?: LegalHoldStatus
   ): Promise<ImageAssetMessageOutgoing> {
     const imageAsset = await this.assetService.uploadImageAsset(image);
 
@@ -164,6 +172,7 @@ class MessageBuilder {
       asset: imageAsset,
       expectsReadConfirmation,
       image,
+      legalHoldStatus,
     };
 
     return {
@@ -180,7 +189,7 @@ class MessageBuilder {
   public createLocation(
     conversationId: string,
     location: LocationContent,
-    messageId: string = MessageBuilder.createId()
+    messageId = MessageBuilder.createId()
   ): LocationMessage {
     return {
       content: location,
@@ -195,14 +204,11 @@ class MessageBuilder {
 
   public createReaction(
     conversationId: string,
-    originalMessageId: string,
-    type: ReactionType,
-    messageId: string = MessageBuilder.createId()
+    reaction: ReactionContent,
+    messageId = MessageBuilder.createId()
   ): ReactionMessage {
-    const content: ReactionContent = {originalMessageId, type};
-
     return {
-      content,
+      content: reaction,
       conversation: conversationId,
       from: this.getSelfUserId(),
       id: messageId,
@@ -212,11 +218,7 @@ class MessageBuilder {
     };
   }
 
-  public createText(
-    conversationId: string,
-    text: string,
-    messageId: string = MessageBuilder.createId()
-  ): TextContentBuilder {
+  public createText(conversationId: string, text: string, messageId = MessageBuilder.createId()): TextContentBuilder {
     const content: TextContent = {text};
 
     const payloadBundle: TextMessage = {
@@ -236,8 +238,8 @@ class MessageBuilder {
     conversationId: string,
     firstMessageId: string,
     type: Confirmation.Type,
-    moreMessageIds?: string[],
-    messageId: string = MessageBuilder.createId()
+    messageId = MessageBuilder.createId(),
+    moreMessageIds?: string[]
   ): ConfirmationMessage {
     const content: ConfirmationContent = {firstMessageId, moreMessageIds, type};
     return {
@@ -254,7 +256,7 @@ class MessageBuilder {
   public createPing(
     conversationId: string,
     ping: KnockContent = {},
-    messageId: string = MessageBuilder.createId()
+    messageId = MessageBuilder.createId()
   ): PingMessage {
     return {
       content: ping,
@@ -267,10 +269,7 @@ class MessageBuilder {
     };
   }
 
-  public createSessionReset(
-    conversationId: string,
-    messageId: string = MessageBuilder.createId()
-  ): ResetSessionMessage {
+  public createSessionReset(conversationId: string, messageId = MessageBuilder.createId()): ResetSessionMessage {
     const content: ClientActionContent = {
       clientAction: ClientAction.RESET_SESSION,
     };