+++ packages/store-engine/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine/compare/@wireapp/store-engine@3.0.1...@wireapp/store-engine@3.0.2) (2019-06-03)
+
+**Note:** Version bump only for package @wireapp/store-engine
+
+
+
+
+
+## [3.0.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine/compare/@wireapp/store-engine@3.0.0...@wireapp/store-engine@3.0.1) (2019-05-31)
+
+**Note:** Version bump only for package @wireapp/store-engine
+
+
+
+
+
 # [3.0.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/store-engine/compare/@wireapp/store-engine@2.1.4...@wireapp/store-engine@3.0.0) (2019-05-28)